+++ api/well-known/webfinger/index.ts
@@ -6,7 +6,6 @@ import { WebFinger } from "@versia/federation/schemas";
 import { User } from "@versia/kit/db";
 import { Users } from "@versia/kit/tables";
 import { and, eq, isNull } from "drizzle-orm";
-import { lookup } from "mime-types";
 import { z } from "zod";
 import { config } from "~/packages/config-manager";
 import { ErrorSchema } from "~/types/api";
@@ -137,9 +136,8 @@ export default apiRoute((app) =>
                     },
                     {
                         rel: "avatar",
-                        // TODO: don't... don't use the mime type from the file extension
                         type:
-                            lookup(user.getAvatarUrl(config)) ||
+                            user.data.source.avatar?.content_type ||
                             "application/octet-stream",
                         href: user.getAvatarUrl(config),
                     },